requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jinja2==3.1.4
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markupsafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
narwhals==2.7.0
networkx==3.3
numpy==2.1.2
openai==2.2.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.0.0
polars==1.34.0
polars-runtime-32==1.34.0
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pydantic==2.12.0
pydantic-core==2.41.1
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scipy==1.16.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
sympy==1.13.3
tenacity==9.1.2
toml==0.10.2
torch==2.8.0+cpu
torchvision==0.23.0+cpu
tornado==6.5.2
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ultralytics==8.3.207
ultralytics-thop==2.0.17
urllib3==2.5.0
watchdog==6.0.0
